+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.1.3
+
   * Fixes to extension parsing #10
   * Better connection/flight timers #9
   * Remove rcgen/ring dependency #8